Vulkan BUILDS!

+++ include/vulkan/ht_vktexture.h
@@ -36,7 +36,7 @@ namespace Hatchit {
                 bool Initialize(Resource::TextureHandle handle, const VkDevice& device);
 
                 //For building a texture *not* from a file
-                bool Initialize(const VkDevice& device, const BYTE* data, size_t width, size_t height, uint32_t channelCount, uint32_t mipLevels);
+                bool Initialize(const VkDevice& device, const BYTE* data, uint32_t width, uint32_t height, uint32_t channelCount, uint32_t mipLevels);
 
                 VkImageView GetView();
